cmd/licenses: add credits-generation tool
There's probably a cuter way to do this, but this is fine. Parse go.mod,
then run each `requir` line through `go mod why -m <module>` to see
who depends on it, then cross-reference that against a manual blacklist
to exclude deps only used in dev. I probably woulddn't have gone to the
trouble of filtering out dev dependencies if not for goreleaser--that
guy's hefty.