cmd/run: better help; include license and OSS credits in binary
I'm slightly annoyed that,
- we have two copies of LICENSE.md, and,
- credits.txt isn't in the repo root

But go:embed doesn't support embedding from parent directoires, and it
doesn't compile without it, and being able to compile off a clone
without doing Opaque Build Steps is very nice.

One alternative is to get rid of the separate cmd/run package, move the
CLI stuff into the root, and rename it to `main`.