*: add "group" task type
If a task has dependencies but no cmd, it should be treated differently
than a regular task:
- it should invalidate tasks that depend on it immediately upon
  being started,
- it should never "exit" (which is to say we shouldn't log about retries
  if Start returns),
- despite not "exit"ing, it should not stop a run from terminating.

In the past, we tried to handle this through synchronized trickery
between scriptTask and Run, such that a Run could detect that a task was
a no-op and treat it specially. This approach made the code hard to
edit, and created footguns for new implementations of the Task
interface.

Here we make no-op-ness an explicit part of TaskMetadata.