*: ui.Start doesn't need IDs, plus fix readme example
I started out just trying to make the README example compile, and then I
noticed that the UIs take IDs effectively twice: once by passing Run to
their constructor, then again as an argument to Start.

If we're passing the whole-ass Run, we don't need to also pass IDs.

Anyway, I also simplified example_test until it was short enough to
paste into the README, and pasted it into the README. This will make it
much easier to ensure it's up-to-date.

We remove the parameter from Start.