integration: exercise printer
The content of each log stream in our tests is predictable, but their
specific interleaving with other streams is not, making it hard to use
in assertions. So, we deinterleave the test streams.

Previously we did this with by using a testUI that never interleaved the
streams in the first place, but that meant,
- we couldn't exercise the real UIs, and,
- the out.log snapshot files didn't look like real output

This commit changes tests to use a printer, saves interleaved output in
out.log snapshots, and only de-interleaves the output and existing
snapshot for the equality check.