wip: real pager with incremental search
Closes #51

bubbles.Viewport isn't really appropriate for a buffer that we keep
appending to like this: we have to manually bookkeep a lot of state back
and forth between the tui and the viewports.

Instead, let's encapsulate the viewports and the log content together
into a new ui component per-stream.

TODO:
- unfuck the ui
- wrap
- hook up the search
- hook up switching in and out of focus
- hook up keys