BREAKING CHANGE: tasks have working directory
Coming back to this repo, it felt weird to me that we pass a dir into
RunTask. In practice, probably the dir has to be right for the task to
work, which means a task isn't well encapsulated: you have to carry both
it -and- its correct dir around.

This change puts the dir into the TaskMetadata.