feat: configurable taskfile name
In another project, I need to have two different taskfiles: one that's
parsed locally, and another that's parsed in Docker. The local file
can't be used in CI because it references subtasks from a folder which
is dockerignored.

It feels kind of gross that you can override the name of the root
taskfile but not sub taskfiles. Maybe there's a more flexible mechanism
for this that I'm not thinking of.