feat: determine runType by root task
Previously, the appearance of any "long" task in a run would make the
whole run "long" (and therefore not exit upon completion of the root
task). That meant you couldn't have, eg, a CI test depend on a
long-running mock server.

Now, we determine the runType based on the root task only.