feat: real pager with incremental search
Closes #51
Closes #77
Closes #105

bubbles.Viewport isn't really appropriate for a buffer that we keep
appending to like this: we have to manually bookkeep a lot of state back
and forth between the tui and the viewports——each time we receive a
line, we send the whole-ass buffer into the viewport.

Instead, let's encapsulate the viewports and the log content together
into a new ui component per-stream.

- adds vim-style incremental search with highlighting
- scroll performance is now roughly O(window height) instead of O(buffer size)
- adds "responsive design", making the ui a lot smaller in small windows
- moves tui help into its own pane, accessed with 'h' or '?'
- adds a bonus package, cmd/scroll, which is a tail replacement of dubious value
- adds some usability features, like a confirmation before quitting

Additionally, we clean up our use of color a bit, so that run looks good
whether you're in dark mode or light.

- make color hash adaptive
- use adaptive colors everywhere
- don't use backgrounds anywhere