build: notarize macos builds in parallel
Waiting for Apple to notarize binaries often takes up the bulk of the
publish script's execution time. Doing the three jobs in parallel ought
to cut that down.

Since this function prints both log output and captured output, we
additionally change the script to print logs to stderr rather than
stdout in general.