feat: only show meta task with fsevents if using fsevents
closes #54

Before this change, we only used this stream for fsevents and the status
at the end of a run. The status events weren't especially useful: it's
weird to keep the UI open after a run is over, and we print exit status
after closing the UI anyway.

This change makes the stream a more purposeful:
- renames it to "watch" (I considered "files", but I like that "watch"
  matches the word used in taskfiles)
- removes the useless status events
- only shows the steram if the run includes file watchers