test: snapshot cli output
color.Render has a crash, but since it's only exercised by a `main`
package, it wasn't caught in tests. This change exercises the `main`
package a bit. It would be nice to do the integration tests this way,
too, but that's a project for another day.

This commit is a failing test; the next commit fixes it.