fix: fix task order again
taskfile ingestion put tasks in order, which was enough to make `run
-list` appear correct, but then when run_task picked the subset for a
given run to put in the sidebar, it put them out-of-order again.

I've now verified that the tasks appear in the correct order in the
sidebar. I ought to write a test for this, but I don't feel like it.

fixes #123